All major bugs fixed (at least i think so)

+++ src/init_wifi.lua
@@ -10,8 +10,11 @@ local print = nil
 if(settings.write_log) then
   print = function(input)
     _print(input)
+    if(input == nil) then
+      input = "nil"
+    end
     local stat = file.stat(".log")
-    if (stat.size > 3000000) then 
+    if (stat and stat.size > 3000000) then 
       file.remove(".log")
     end
     log_file = file.open(".log", "a+")
